% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf.r
\name{ddf}
\alias{ddf}
\title{Distance Detection Function Fitting}
\usage{
ddf(
  dsmodel = call(),
  mrmodel = call(),
  data,
  method = "ds",
  meta.data = list(),
  control = list(),
  call = NULL
)
}
\arguments{
\item{dsmodel}{distance sampling model specification}

\item{mrmodel}{mark-recapture model specification}

\item{data}{dataframe containing data to be analyzed}

\item{method}{analysis method}

\item{meta.data}{list containing settings controlling data structure}

\item{control}{list containing settings controlling model fitting}

\item{call}{not implemented for top level ddf function, this is set by ddf as it is passed to the other ddf generics.}
}
\value{
model object of class=(method, "ddf")
}
\description{
Generic function for fitting detection functions for distance sampling with
single and double observer configurations. Independent observer, trial and
dependent observer (removal) configurations are included. This is a generic
function which does little other than to validate the calling arguments and
methods and then calls the appropriate \code{method} specific function to do
the analysis. This is a temp version to replace what is in mrds. Not all functionality
is included here.
}
\author{
Jeff Laake
# load data
}
