% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.r
\name{delta}
\alias{delta}
\title{Delta computation for Buckland et al 2010 model}
\usage{
delta(x,p1,p2,gamma,PI=FALSE,use.offset=FALSE)
}
\arguments{
\item{x}{a design matrix for delta}

\item{p1}{detection probabilities for observer 1}

\item{p2}{detection probabilities for observer 2}

\item{PI}{TRUE if this is a point independence model}

\item{use.offset}{if TRUE use offset value for logit}
}
\value{
Vector of delta values for each row in design matrix x
}
\description{
Computes delta values for all values in design matrix, See Buckland et al, (2010) for
definition
}
