% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen,history.r
\name{gen.history}
\alias{gen.history}
\title{Generate a set of double observer detection histories from population data and model}
\usage{
gen.history(x,par,p.formula,delta.formula,indep=FALSE,PI=FALSE,debug=FALSE,use.offset=FALSE)
}
\arguments{
\item{x}{population data - 1 row per observer and must contain and observer and object fields at a minimum}

\item{par}{model parameter values}

\item{p.formula}{formula for true detection probabilities}

\item{delta.formula}{formula for delta dependence function}

\item{indep}{if TRUE, uses full independence model}

\item{PI}{if TRUE, uses point independence model}

\item{debug}{plots p1, conditional p1 and delta values}

\item{use.offset}{if TRUE use offset value for logit}
}
\value{
simulation observation dataframe;The return value is the dataframe of the observations
seen by at least one observer and the field detected is added to indicate whether the observer
 detected the observation.  Those missed by both observers are not included.
}
\description{
The return value is the dataframe of the observations seen by at least
one observer and the field detected is added to indicate whether the observer
detected the observation.  Those missed by both observers are not included.
}
