% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.general.R
\name{halfnormal}
\alias{halfnormal}
\title{Half-normal detection function}
\usage{
halfnormal(distance, scale)
}
\arguments{
\item{distance}{vector of distances}

\item{scale}{vector of scale values on log scale for sigma}
}
\value{
vector of detection probabilities
}
\description{
Half-normal detection function
}
\author{
Jeff Laake
}
