% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_mu.bpi.R
\name{integrate_mu.bpi}
\alias{integrate_mu.bpi}
\title{Integral of probability of detection by at least one observer for fitted bpi model}
\usage{
integrate_mu.bpi(
  x,
  n,
  pformula,
  dformula,
  par,
  width,
  indep,
  PI,
  use.offset,
  posdep
)
}
\arguments{
\item{x}{observation dataframe}

\item{n}{number of integrals to compute}

\item{pformula}{formula for detection probabilities}

\item{dformula}{formula for delta dependence function}

\item{par}{parameter values}

\item{width}{transect half-width}

\item{indep}{if TRUE, uses full independence model}

\item{PI}{if TRUE, uses point independence model}

\item{use.offset}{if TRUE use offset value for logit}

\item{posdep}{if TRUE forces positive dependence relationship}
}
\value{
vector of n integral values
}
\description{
Calls \link{mu.bpi} to compute integrals over detection function values for
each observation to integrate average detection probability by at least one observer.
}
\author{
Jeff Laake
}
