% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_mu.loglinear.R
\name{integrate_mu.loglinear}
\alias{integrate_mu.loglinear}
\title{Integral of probability of detection by at least one observer for fitted loglinear model}
\usage{
integrate_mu.loglinear(x, pformula, dformula, par, width)
}
\arguments{
\item{x}{observation dataframe}

\item{pformula}{formula for detection probabilities}

\item{dformula}{formula for delta dependence function}

\item{par}{parameters for model}

\item{width}{transect half-width}

\item{n}{number of integrals to compute}
}
\value{
vector of n integral values
}
\description{
Calls \link{mu.loglinear} to compute integrals over detection function values for
each observation to integrate average detection probability by at least one observer.
}
\author{
Jeff Laake
}
