% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.r
\name{integratelogistic}
\alias{integratelogistic}
\title{Computes integral of logistic function from 0 to width}
\usage{
integratelogistic(x1, x2, models, beta, gamma, width,indep,PI,use.offset)
}
\arguments{
\item{x1}{covariate data for observer1}

\item{x2}{covariate data for observer2}

\item{models}{model list of p and delta formulas for detection probabilities}

\item{beta}{parameters for p}

\item{gamma}{parameters for delta dependence function}

\item{width}{transect half-width}

\item{indep}{if TRUE, uses full independence model}

\item{PI}{if TRUE, uses point independence model}

\item{use.offset}{if TRUE use offset value for logit}
}
\value{
integral value
}
\description{
Uses logisticbyx to compute integral from 0 to width for a specific set of covariate values
}
