% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.bpi.R
\name{ll.bpi}
\alias{ll.bpi}
\title{Log likelihood calculation for Buckland et al 2010 model}
\usage{
ll.bpi(
  par,
  x,
  p.formula,
  delta.formula,
  width,
  debug = FALSE,
  indep = FALSE,
  PI = FALSE,
  use.offset = FALSE,
  posdep = FALSE
)
}
\arguments{
\item{par}{model parameter values}

\item{x}{observation dataframe}

\item{p.formula}{formula for detection probabilities}

\item{delta.formula}{formula for delta dependence function}

\item{width}{transect half-width}

\item{debug}{if TRUE, outputs parameter and log likelihood values}

\item{indep}{if TRUE, uses full independence model}

\item{PI}{if TRUE, uses point independence model}

\item{use.offset}{if TRUE use offset value for logit}

\item{posdep}{if TRUE enforces positive dependence in delta}
}
\value{
negative log-likelihood value at values of par
}
\description{
For a set of observations, computes observer specific probabilities, integrals and delta values See Buckland et al, (2010) for
definitions
}
