% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.mr.R
\name{ll.mr}
\alias{ll.mr}
\title{Log likelihood calculation for mark-recapture type Huggins model}
\usage{
ll.mr(par, x, pformula, dformula, width, debug, indep)
}
\arguments{
\item{par}{model parameter values}

\item{x}{observation dataframe}

\item{pformula}{formula for detection probabilities}

\item{dformula}{formula for delta dependence function}

\item{width}{transect half-width}

\item{debug}{if TRUE, outputs parameter and log likelihood values}

\item{indep}{if TRUE, uses full independence model}
}
\value{
negative log-likelihood value at values of par
}
\description{
Log likelihood calculation for mark-recapture type Huggins model
}
