% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.r
\name{logisticbyx}
\alias{logisticbyx}
\title{Computes values of logistic at a range of x for integration}
\usage{
logisticbyx(distance, x1, x2, models, beta, gamma,indep=FALSE,PI=FALSE,use.offset=FALSE)
}
\arguments{
\item{distance}{vector of distance values}

\item{x1}{covariate data for observer1}

\item{x2}{covariate data for observer2}

\item{models}{model list ofp and delta formulas for detection probabilities}

\item{beta}{parameters for p}

\item{gamma}{prameters for delta dependence function}

\item{indep}{if TRUE, uses full independence model}

\item{PI}{if TRUE, uses point independence model}

\item{use.offset}{if TRUE use offset value for logit}
}
\value{
vector of values of logistic with specific covariate values over a set of distances
}
\description{
For a given z it computes logistic function with those covariate values for a set of distances
}
