% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.design.data.R
\name{make.design.data}
\alias{make.design.data}
\title{Create a dataframe with 2 records (one fore each observer) for each observation from a dataframe with a single record per observation contianing a capture history (ch), distance and
any covariate values.}
\usage{
make.design.data(x, persons = NULL)
}
\arguments{
\item{x}{a dataframe with a single record per observation}

\item{persons}{a dataframe containing a single record per observation and 2 columns indicating which person was in observer (position) 1 and 2 which provides the opportunity to estimate of personnel effects on detection}
}
\value{
a dataframe with 2 records for each observation (format of mrds package)
}
\description{
Create a dataframe with 2 records (one fore each observer) for each observation from a dataframe with a single record per observation contianing a capture history (ch), distance and
any covariate values.
}
\author{
Jeff Laake
}
