% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.bpi.R
\name{mu.bpi}
\alias{mu.bpi}
\title{Integral of probability of detection by at least one observer for fitted bpi model}
\usage{
mu.bpi(
  x,
  n,
  p.formula,
  delta.formula,
  beta,
  gamma,
  width,
  indep,
  PI,
  use.offset
)
}
\arguments{
\item{x}{observation dataframe}

\item{n}{number of integrals to compute}

\item{p.formula}{formula for detection probabilities}

\item{delta.formula}{formula for delta dependence function}

\item{beta}{parameters for p.formula}

\item{gamma}{parameters for delta.formula}

\item{width}{transect half-width}

\item{indep}{if TRUE, uses full independence model}

\item{PI}{if TRUE, uses point independence model}

\item{use.offset}{if TRUE use offset value for logit}
}
\value{
vector of n integral values
}
\description{
Calls \link{integratelogistic} to compute integrals over detection function values for
each observation to integrate average detection probability by at least one observer.
}
\author{
Jeff Laake
}
