% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.loglinear.R
\name{mu.loglinear}
\alias{mu.loglinear}
\title{Integral of probability of detection by at least one observer for fitted log-linear model}
\usage{
mu.loglinear(x, par, dd, pformula, dformula)
}
\arguments{
\item{x}{vector of distance values from integrate to compute integral}

\item{par}{parameter starting values; if NULL they are computed}

\item{dd}{a single pair of detection records for the integration; the pair is replicated length(x) times to compute probability at each x for the integration}

\item{pformula}{formula for detection but excluding the dependence portion}

\item{dformula}{model for dependence due residual (unmodelled) heterogeneity}
}
\value{
vector of values at different x (distance) values for integrate
}
\description{
Calls \link{p.loglinear} to compute 1-1/K for a range of x (distance) values for
each observation to integrate average detection probability by at least one observer.
}
\seealso{
\link{p.loglinear}
}
\author{
Jeff Laake
}
