% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.loglinear.R
\name{p.loglinear}
\alias{p.loglinear}
\title{Compute detection probabilities for each observable outcome (10,01,11) and K for log-linear model with possible dependence in detections for mark-recapture distance data
data.}
\usage{
p.loglinear(par, x, pformula, dformula)
}
\arguments{
\item{par}{parameter values}

\item{x}{dataframe containing double observer data (2 records per observation) with fields observer, detected, distance, ch and any covariate values}

\item{pformula}{formula for detection but excluding the dependence portion}

\item{dformula}{model for dependence due to residual (unmodelled) heterogeneity}
}
\value{
list with prob: probability matrix with 3 columns for 10,01,11 at each of n rows of observations and K the denominator
}
\description{
Compute detection probabilities for each observable outcome (10,01,11) and K for log-linear model with possible dependence in detections for mark-recapture distance data
data.
}
\author{
Jeff Laake
}
