% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_det.R
\name{plot_det}
\alias{plot_det}
\title{Plot some detection function values. This is very simple at present.}
\usage{
plot_det(mod)
}
\arguments{
\item{mod}{fitted result from fit_mrdsalt}
}
\value{
None
}
\description{
Plot some detection function values. This is very simple at present.
}
\author{
Jeff Laake
}
