% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bpi.R
\name{predict.bpi}
\alias{predict.bpi}
\title{Compute average detection probabilities for Buckland et al model with possible dependence in detections for mark-recapture distance data.}
\usage{
\method{predict}{bpi}(object, ...)
}
\arguments{
\item{object}{fitted bpi model}

\item{...}{additional arguments affecting the predictions produced}
}
\value{
vector of predicted average detection probabilities (integrated over x) for each observation.
}
\description{
Compute average detection probabilities for Buckland et al model with possible dependence in detections for mark-recapture distance data.
}
\seealso{
\link{integrate_mu.bpi}
}
\author{
Jeff Laake
}
