% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mr.R
\name{predict.mr}
\alias{predict.mr}
\title{Compute average detection probabilities for mark-recapture ty with possible dependence in detections for mark-recapture distance data.}
\usage{
\method{predict}{mr}(object, ...)
}
\arguments{
\item{object}{fitted mr model}

\item{...}{additional arguments affecting the predictions produced}
}
\value{
vector of predicted average detection probabilities (integrated over x) for each observation.
}
\description{
Compute average detection probabilities for mark-recapture ty with possible dependence in detections for mark-recapture distance data.
}
\seealso{
\link{integrate_mu.mr}
}
\author{
Jeff Laake
}
