% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.loglinear.R
\name{sim.loglinear}
\alias{sim.loglinear}
\title{Simulates some mark-recapture distance data using a log-linear model with possible dependence in detections.
data.}
\usage{
sim.loglinear(
  N,
  W,
  par,
  cov = NULL,
  pformula = ~-1 + observer + observer:distance,
  dformula = ~-1 + distance
)
}
\arguments{
\item{N}{abundance of population which is sampled by dual observers}

\item{W}{half-width of the transect; observations are out to distance W.}

\item{par}{parameter values}

\item{cov}{a function that creates a single covariate value named cov to be used in the data simulation}

\item{pformula}{formula for detection but excluding the dependence portion}

\item{dformula}{model for dependence due residual (unmodelled) heterogeneity}
}
\value{
dataframe containing a pair of records (one for each observer) with the fields to be used for analysis including the capture history. Objects not seen ch="00" are excluded from the dataframe.
}
\description{
Simulates some mark-recapture distance data using a log-linear model with possible dependence in detections.
data.
}
\author{
Jeff Laake
}
