% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bpi.r
\name{summary.bpi}
\alias{summary.bpi}
\title{Summary of bpi model fit}
\usage{
\method{summary}{bpi}(object,se=TRUE,...)
}
\arguments{
\item{object}{a ddf model object}

\item{se}{if TRUE, computes std errors}

\item{...}{additional non-specified argument for S3 generic function}
}
\value{
A list with each of the summarized objects that depends on the model
}
\description{
Creates a summary object of a ddf loglinear model fitted to double observer data in
an independent observer configuration
}
