% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.bpi.R
\name{mu.bpi}
\alias{mu.bpi}
\title{Computes probability of detection by at least one observer for fitted bpi model at a range of distances to
integrate pooled detection function}
\usage{
mu.bpi(x, dd, models, par, indep, PI, use.offset, posdep)
}
\arguments{
\item{x}{vector of distance values}

\item{dd}{a single pair of detection records for the integration; the pair is replicated length(x) times to compute probability at each x for the integration}

\item{models}{model list of p and delta formulas for detection probabilities}

\item{indep}{if TRUE, uses full independence model}

\item{PI}{if TRUE, uses point independence model}

\item{use.offset}{if TRUE use offset value for logit}

\item{posdep}{if TRUE enforces positive dependence in delta}

\item{beta}{parameters for p}

\item{gamma}{parameters for delta dependence function}
}
\value{
vector of values of probability of detection by at least one observer for fitted bpi model at a set of distance
}
\description{
Computes probability of detection by at least one observer for fitted bpi model at a range of distances to
integrate pooled detection function
}
