% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.mr.r
\name{p.mr}
\alias{p.mr}
\title{Probability calculations for Buckland et al 2010 model}
\usage{
p.mr(par, x, pformula, dformula, indep)
}
\arguments{
\item{par}{model parameter values}

\item{x}{observation dataframe}

\item{pformula}{formula for detection probabilities}

\item{dformula}{formula for delta dependence function}

\item{indep}{if TRUE, uses full independence model}

\item{PI}{if TRUE, uses point independence model}

\item{use.offset}{if TRUE use offset value for logit}

\item{posdep}{if TRUE enforces positive dependence in delta}
}
\value{
list of vectors for p1,p2, integrals and delta values
}
\description{
For a set of observations, computes observer specific probabilities, integrals and delta values See Buckland et al, (2010) for
definitions
}
