% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.loglinear.R
\name{predict.loglinear}
\alias{predict.loglinear}
\title{Compute average detection probabilities for log-linear model with possible dependence in detections for mark-recapture distance data.}
\usage{
\method{predict}{loglinear}(object, ...)
}
\arguments{
\item{object}{fitted loglinear model}

\item{...}{additional arguments affecting the predictions produced}
}
\value{
vector of predicted average detection probabilities (integrated over x) for each observation.
}
\description{
Compute average detection probabilities for log-linear model with possible dependence in detections for mark-recapture distance data.
}
\seealso{
\link{integrate_mu.loglinear}
}
\author{
Jeff Laake
}
